+++ src/tbb/tbb_misc.h
@@ -150,8 +150,8 @@ class ExponentialBackoff {
     //! Time delay, in units of "pause" instructions. 
     /** Should be equal to approximately the number of "pause" instructions
         that take the same time as an context switch. */
-    static const uintptr LOOPS_BEFORE_YIELD = 0x10;
-    uintptr count;
+    static const int LOOPS_BEFORE_YIELD = 0x10;
+    int count;
 public:
     ExponentialBackoff() : count(1) {}
